Removing old formatter in ray.h
