+++ include/xdg/ray.h
@@ -7,8 +7,6 @@
 
 #include "xdg/vec3da.h"
 
-#include <fmt/format.h>
-
 #include "xdg/constants.h"
 #include "xdg/embree_interface.h"
 
@@ -148,18 +146,5 @@ struct RTCDPointQuery : RTCPointQuery {
 
 } // namespace xdg
 
-namespace fmt {
-
-template<>
-struct formatter<xdg::Vec3da> : formatter<std::string> {
-  template<typename FormatContext>
-  auto format(const xdg::Vec3da& v, FormatContext& ctx)
-  {
-    return formatter<std::string>::format(
-      fmt::format("({}, {}, {})", v.x, v.y, v.z), ctx);
-  }
-};
-
-} // namespace fmt
 
 #endif // include guard
