+++ include/xdg/vec3da.h
@@ -17,6 +17,8 @@
 #define __forceinline inline __attribute__((always_inline))
 #endif
 
+#include <fmt/format.h>
+
 #include "xdg/constants.h"
 
 namespace xdg {
@@ -175,18 +177,15 @@ inline Direction rand_dir() {
 
 } // end namespace xdg
 
-namespace fmt{
-template <>
-struct formatter<xdg::Vec3da> {
-    template <typename ParseContext>
-    constexpr auto parse(ParseContext& ctx) {
-        return ctx.begin();
-    }
-
-    template <typename FormatContext>
-    auto format(const xdg::Vec3da& v, FormatContext& ctx) const {
-        return fmt::format_to(ctx.out(), "[{}, {}, {}]", v.x, v.y, v.z);
-    }
+namespace fmt {
+template<>
+struct formatter<xdg::Vec3da> : formatter<std::string> {
+  template<typename FormatContext>
+  auto format(const xdg::Vec3da& v, FormatContext& ctx)
+  {
+    return formatter<std::string>::format(
+      fmt::format("({}, {}, {})", v.x, v.y, v.z), ctx);
+  }
 };
 
 } // end namespace fmt
